clean up tag summary state
